vibz/pyproject.toml
[project]
name = "vibz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.0",
    "gitpython>=3.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]
